+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 * Issue with slurm 21.x
 
+### Added
+* Experimental feature to automatically handle activating virtual environments, currently disabled by default set  `--virtualenv_cmd=auto` to use with cmdqueue boilerplate scripts.
+
 
 ## Version 0.2.2 - Released 2025-02-19